+++ packages/cinterop/src/jvmCommon/kotlin/io/realm/interop/RealmInterop.kt
@@ -120,8 +120,24 @@ actual object RealmInterop {
 
     actual fun realm_open(config: NativePointer): NativePointer {
         val realmPtr = LongPointerWrapper(realmc.realm_open((config as LongPointerWrapper).ptr))
+        // Ensure that we can read version information, etc.
         realm_begin_read(realmPtr)
         return realmPtr
+    }
+
+    actual fun realm_freeze(liveRealm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_freeze(liveRealm.cptr()))
+    }
+
+    actual fun realm_thaw(frozenRealm: NativePointer): NativePointer {
+        val realmPtr = LongPointerWrapper(realmc.realm_thaw(frozenRealm.cptr()))
+        // Ensure that we can read version information, etc.
+        realm_begin_read(realmPtr)
+        return realmPtr
+    }
+
+    actual fun realm_is_frozen(realm: NativePointer): Boolean {
+        return realmc.realm_is_frozen(realm.cptr())
     }
 
     actual fun realm_close(realm: NativePointer) {
@@ -157,6 +173,10 @@ actual object RealmInterop {
         realmc.realm_begin_write((realm as LongPointerWrapper).ptr)
     }
 
+    actual fun realm_is_in_transaction(realm: NativePointer): Boolean {
+        return realmc.realm_is_writable(realm.cptr())
+    }
+
     actual fun realm_commit(realm: NativePointer) {
         realmc.realm_commit((realm as LongPointerWrapper).ptr)
     }
@@ -168,8 +188,21 @@ actual object RealmInterop {
     actual fun realm_object_create(realm: NativePointer, key: Long): NativePointer {
         return LongPointerWrapper(realmc.realm_object_create((realm as LongPointerWrapper).ptr, key))
     }
+
     actual fun realm_object_create_with_primary_key(realm: NativePointer, key: Long, primaryKey: Any?): NativePointer {
         return LongPointerWrapper(realmc.realm_object_create_with_primary_key((realm as LongPointerWrapper).ptr, key, to_realm_value(primaryKey)))
+    }
+
+    actual fun realm_object_is_valid(obj: NativePointer): Boolean {
+        return realmc.realm_object_is_valid(obj.cptr())
+    }
+
+    actual fun realm_object_freeze(liveObject: NativePointer, frozenRealm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_object_freeze(liveObject.cptr(), frozenRealm.cptr()))
+    }
+
+    actual fun realm_object_thaw(frozenObject: NativePointer, liveRealm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_object_thaw(frozenObject.cptr(), liveRealm.cptr()))
     }
 
     actual fun realm_find_class(realm: NativePointer, name: String): Long {
@@ -353,6 +386,14 @@ actual object RealmInterop {
 
     actual fun realm_query_find_all(query: NativePointer): NativePointer {
         return LongPointerWrapper(realmc.realm_query_find_all(query.cptr()))
+    }
+
+    actual fun realm_results_freeze(liveResults: NativePointer, frozenRealm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_results_freeze(liveResults.cptr(), frozenRealm.cptr()))
+    }
+
+    actual fun realm_results_thaw(frozenResults: NativePointer, liveRealm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_results_thaw(frozenResults.cptr(), liveRealm.cptr()))
     }
 
     actual fun realm_results_count(results: NativePointer): Long {